ok it wors

+++ Arduino/inno_22/src/pmath.h
@@ -94,7 +94,7 @@ namespace pmath
         Vector orth();
         
     private:
-        int x, y;
+        long x, y;
     };    
 }
